# Optimise alignment/part7 vectorised banded alignment