--- a/test/performance/alignment/CMakeLists.txt
+++ b/test/performance/alignment/CMakeLists.txt
@@ -1,3 +1,4 @@
+seqan3_benchmark(global_affine_alignment_banded_simd_benchmark.cpp)
 seqan3_benchmark(global_affine_alignment_benchmark.cpp)
 seqan3_benchmark(global_affine_alignment_parallel_benchmark.cpp)
 seqan3_benchmark(global_affine_alignment_protein_simd_benchmark.cpp)