… test/unit/alignment/pairwise/CMakeLists.txt