Initial commit to run vectorised banded alignment. This change includes the basic setup to run, test and measure the performance of the vectorised banded version of the alignment. 

seqan/product_backlog#54

Here the current benchmark:

```console
-------------------------------------------------------------------------------------------------------------------------
Benchmark                                                               Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------------------------------
seqan3_affine_accelerated/simd_with_score/0/real_time              983204 ns       976844 ns          712 CUPS=3.18801G/s cells=3.13446M total=-70.5478M
seqan3_affine_accelerated/simd_parallel_with_score/0/real_time     732206 ns       350839 ns          946 CUPS=4.28085G/s cells=3.13446M total=-93.7335M
seqan2_affine_accelerated/simd_with_score/0/real_time              646466 ns       637440 ns         1114 CUPS=4.84861G/s cells=3.13446M total=-110.38M
seqan2_affine_accelerated/simd_parallel_with_score/0/real_time     639347 ns       633055 ns         1052 CUPS=4.9026G/s cells=3.13446M total=-104.236M
```